// UI/src/app/Apis/ControllerPaths/ControllerPaths.ts
export enum CRMControllerPaths {
  PROJECT_CHARGE = '/projectcharge',
  CUSTOMER_STAGE = '/customerstages',
  APPLICANT_INFO = '/applicantinfo',
  BOOKING_CHARGE = '/bookingcharges',
  PAYMENT_DETAILS = '/paymentdetails',
  CUSTOMER_PAYMENTS = '/customerpayment',
  DASH_BOARD = '/dashboard',
  CRM_DOCUMENTS = '/crm/documents',
  PROJECT_DETAILS = '/projectdetails',
  CRM_FOLLOWUP = '/crm/followup',
}